Updated data field to now be a decimal

+++ DjangoProject/SensorPoll/models.py
@@ -3,7 +3,7 @@
 class SensorPacket(models.Model):
 	address = models.IntegerField()
 	time_stamp = models.DateTimeField(auto_now_add=True)
-	data = models.IntegerField()
+	data = models.DecimalField(max_digits=6, decimal_places=2)
 
 class AveragedData(models.Model):
     address = models.IntegerField()
